requirements.txt
Django==2.2.3
django-rest==0.0.1
djangorestframework==3.10.1
envparse==0.2.0
numpy==1.16.4
pandas==0.25.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
